backend/requirements.txt
fastapi==0.136.1
httpx==0.28.1
pydantic==2.13.3
pytest==9.0.3
uvicorn[standard]==0.46.0
pdfplumber>=0.11.0
pyjwt[crypto]>=2.8.0
cryptography>=42.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
